Fix Settings link to show for symlinked plugin dir (#5)
When this plugin was symlinked into the plugins dir from a directory
named something other than the plugin slug, the Settings link was
not shown for this plugin in the Plugins list. That is now fixed.

This also adds a capabilities check, so the link is not shown to people
without permission to change the settings.