Strengthen cache dir checks
Since mkdir() is affected by the current umask, it is possible for a
directory to be created without write or execute permissions. This
update moves the cache writability check to after we attempt to create
the caching directory so we can confirm that it truly succeeded and is
writable and executable.