Cache cleanup with WP Cron
Clear cache once per day. Remove all files older
than 24 hours. They will get regenerated on next request.

No need to check inside service.php for file age. We
send the filemtime (which is max file mtime of the concated stuff)
as part of the request URI, which is
hashed to generate the cache file. If none of the files change,
we shouldn't do anything.

Update naming pattern, so we can "reliably" cleanup.