Limit cached headers
Prior to this change, we cached all headers sent before the concat
service responds. In most cases, that would be fine because the service
responds very early to the request, but it is technically possible for
another sensitive header to be added prior to response and cached
for sharing with other requestors.

This change removes that possibility by only caching a specific set of
headers.