Revert original change to abspath-based resolution
It doesn't fix the doubled path separator issues in all cases.
That issue can be fixed in a separate PR. It doesn't appear to be
a problem in practice but mucks with tests when you want to assert
that paths are equal. And it would be good to return clean paths anyway.