Reduce interference with script tag filtering (#45)
When we encounter a single-script concat group, we aren't actually
concatenating, yet because the script is expected to be concatenated,
we handle printing the script rather than relying on the do_item()
method. This means that the core `script_loader_tag` filter is not
applied to the script tag even though it is still a normal,
single-script tag.

We recently discovered that this interfere's with
the TwentyTwenty theme adding async/defer attributes to its scripts
even though they are not being concatenated with anything else.

We address this by manually applying the `script_loader_tag` filter
when we are printing the script tag for a single-script concat group.